+++ golf-sync-swing/Services/SwingDetectorProtocol.swift
@@ -32,5 +32,4 @@ protocol RealTimeSwingDetector: AnyObject, Sendable {
 
 // MARK: - Protocol Conformance
 
-extension SwingNetDetector: RealTimeSwingDetector {}
 extension ActionClassifierDetector: RealTimeSwingDetector {}